commit halpfay through Help derive implementation, will refactor to reflect changes in jagaimo_macro
