+++ jagaimo_derive_help/Cargo.toml
@@ -15,3 +15,4 @@ proc-macro2 = "1.0.95"
 quote = "1.0.40"
 syn = { version = "2.0.101", features = ["extra-traits"] }
 toml = "0.8.23"
+resolve_calling_crate = { path = "../resolve_calling_crate/" }
